requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
paho-mqtt==2.1.0
streamlit==1.34.0
pandas==2.2.2
requests==2.32.2
httpx==0.25.1
numpy==1.26.4
scikit-learn==1.4.0
lightgbm==4.1.0
joblib==1.3.2
